refactor(common): Annotation, Aspect Common으로 이동
