+++ com.delivery-signal.eureka.client.hub/src/main/java/com/delivery_signal/eureka/client/hub/common/annotation/CacheEvict.java
@@ -1,4 +1,4 @@
-package com.delivery_signal.eureka.client.hub.infrastructure.annotation;
+package com.delivery_signal.eureka.client.hub.common.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,13 +6,11 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import org.springframework.stereotype.Component;
-
-@Target(ElementType.TYPE)
+@Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@Component
-public @interface ExternalAdapter {
-
-	String value() default "";
+public @interface CacheEvict {
+	String value();
+	String key() default "";
+	boolean allEntries() default false;
 }
