+++ com.delivery-signal.eureka.client.hub/src/main/java/com/delivery_signal/eureka/client/hub/common/annotation/ExternalAdapter.java
@@ -1,4 +1,4 @@
-package com.delivery_signal.eureka.client.hub.infrastructure.annotation;
+package com.delivery_signal.eureka.client.hub.common.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -6,11 +6,13 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target(ElementType.METHOD)
+import org.springframework.stereotype.Component;
+
+@Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface CacheEvict {
-	String value();
-	String key() default "";
-	boolean allEntries() default false;
+@Component
+public @interface ExternalAdapter {
+
+	String value() default "";
 }
